src/openforms/prefill/tests/test_registry.py
from django.test import SimpleTestCase

from ..base import BasePlugin
from ..registry import Registry


class Plugin(BasePlugin):
    verbose_name = "some human readable label"


class RegistryTests(SimpleTestCase):
    def tset_register_function(self):
        register = Registry()

        register("plugin1")(Plugin)

        plugin = register["plugin1"]

        self.assertIsInstance(plugin, Plugin)
        self.assertEqual(plugin.identifier, "plugin1")
        self.assertEqual(plugin.verbose_name, "some human readable label")

    def test_duplicate_identifier(self):
        register = Registry()
        register("plugin")(Plugin)

        with self.assertRaisesMessage(
            ValueError,
            "The unique identifier 'plugin' is already present in the registry",
        ):
            register("plugin")(Plugin)
